src/main/java/com/example/recipeapp/controllers/IndexController.java
package com.example.recipeapp.controllers;

import com.example.recipeapp.domain.Recipe;
import com.example.recipeapp.service.RecipeService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;

import java.util.Comparator;
import java.util.List;
import java.util.stream.Collectors;

@Slf4j
@Controller
public class IndexController {

    private final RecipeService recipeService;

    public IndexController(RecipeService recipeService) {
        this.recipeService = recipeService;
    }

    @RequestMapping({"", "/", "/index"})
    public String getIndexPage(Model model) {
        log.debug("Getting Index page");

        List<Recipe> recipeSortedById = recipeService.getRecipes()
            .stream()
            .sorted(Comparator.comparing(Recipe::getId))
            .collect(Collectors.toList());

        model.addAttribute("recipes", recipeSortedById);

        return "index";
    }
}
